Refactor JT Graph
