+++ src/graphs/index.ts
@@ -1,3 +1,4 @@
 export * from './graph';
 export * from './moralGraph';
 export * from './triangulatedGraph';
+export * from './junctionTreeGraph';
